refactor: correct naming of structs and enums
