+++ src/api/resources/auth_key_resource.rs
@@ -6,13 +6,13 @@ use crate::protocol::clock::DurationSinceUnixEpoch;
 use crate::tracker::key::Auth;
 
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
-pub struct AuthKeyResource {
+pub struct AuthKey {
     pub key: String,
     pub valid_until: Option<u64>,
 }
 
-impl From<AuthKeyResource> for Auth {
-    fn from(auth_key_resource: AuthKeyResource) -> Self {
+impl From<AuthKey> for Auth {
+    fn from(auth_key_resource: AuthKey) -> Self {
         Auth {
             key: auth_key_resource.key,
             valid_until: auth_key_resource
@@ -22,9 +22,9 @@ impl From<AuthKeyResource> for Auth {
     }
 }
 
-impl From<Auth> for AuthKeyResource {
+impl From<Auth> for AuthKey {
     fn from(auth_key: Auth) -> Self {
-        AuthKeyResource {
+        AuthKey {
             key: auth_key.key,
             valid_until: auth_key.valid_until.map(|valid_until| valid_until.as_secs()),
         }
@@ -35,7 +35,7 @@ impl From<Auth> for AuthKeyResource {
 mod tests {
     use std::time::Duration;
 
-    use super::AuthKeyResource;
+    use super::AuthKey;
     use crate::protocol::clock::{Current, TimeNow};
     use crate::tracker::key::Auth;
 
@@ -43,7 +43,7 @@ mod tests {
     fn it_should_be_convertible_into_an_auth_key() {
         let duration_in_secs = 60;
 
-        let auth_key_resource = AuthKeyResource {
+        let auth_key_resource = AuthKey {
             key: "IaWDneuFNZi8IB4MPA3qW1CD0M30EZSM".to_string(), // cspell:disable-line
             valid_until: Some(duration_in_secs),
         };
@@ -67,8 +67,8 @@ mod tests {
         };
 
         assert_eq!(
-            AuthKeyResource::from(auth_key),
-            AuthKeyResource {
+            AuthKey::from(auth_key),
+            AuthKey {
                 key: "IaWDneuFNZi8IB4MPA3qW1CD0M30EZSM".to_string(), // cspell:disable-line
                 valid_until: Some(duration_in_secs)
             }
@@ -78,7 +78,7 @@ mod tests {
     #[test]
     fn it_should_be_convertible_into_json() {
         assert_eq!(
-            serde_json::to_string(&AuthKeyResource {
+            serde_json::to_string(&AuthKey {
                 key: "IaWDneuFNZi8IB4MPA3qW1CD0M30EZSM".to_string(), // cspell:disable-line
                 valid_until: Some(60)
             })
