+++ src/protocol/crypto.rs
@@ -1,18 +1,18 @@
 pub mod keys {
 
     pub mod seeds {
-        use self::detail::DEFAULT_SEED;
+        use self::detail::CURRENT_SEED;
         use crate::ephemeral_instance_keys::{Seed, RANDOM_SEED};
 
-        pub trait SeedKeeper {
+        pub trait Keeper {
             type Seed: Sized + Default + AsMut<[u8]>;
             fn get_seed() -> &'static Self::Seed;
         }
 
-        pub struct InstanceSeed;
-        pub struct DefaultSeed;
+        pub struct Instance;
+        pub struct Current;
 
-        impl SeedKeeper for InstanceSeed {
+        impl Keeper for Instance {
             type Seed = Seed;
 
             fn get_seed() -> &'static Self::Seed {
@@ -20,24 +20,24 @@ pub mod keys {
             }
         }
 
-        impl SeedKeeper for DefaultSeed {
+        impl Keeper for Current {
             type Seed = Seed;
 
             #[allow(clippy::needless_borrow)]
             fn get_seed() -> &'static Self::Seed {
-                &DEFAULT_SEED
+                &CURRENT_SEED
             }
         }
 
         #[cfg(test)]
         mod tests {
             use super::detail::ZEROED_TEST_SEED;
-            use super::{DefaultSeed, InstanceSeed, SeedKeeper};
+            use super::{Current, Instance, Keeper};
             use crate::ephemeral_instance_keys::Seed;
 
             pub struct ZeroedTestSeed;
 
-            impl SeedKeeper for ZeroedTestSeed {
+            impl Keeper for ZeroedTestSeed {
                 type Seed = Seed;
 
                 #[allow(clippy::needless_borrow)]
@@ -48,12 +48,12 @@ pub mod keys {
 
             #[test]
             fn the_default_seed_and_the_zeroed_seed_should_be_the_same_when_testing() {
-                assert_eq!(DefaultSeed::get_seed(), ZeroedTestSeed::get_seed());
+                assert_eq!(Current::get_seed(), ZeroedTestSeed::get_seed());
             }
 
             #[test]
             fn the_default_seed_and_the_instance_seed_should_be_different_when_testing() {
-                assert_ne!(DefaultSeed::get_seed(), InstanceSeed::get_seed());
+                assert_ne!(Current::get_seed(), Instance::get_seed());
             }
         }
 
@@ -64,10 +64,10 @@ pub mod keys {
             pub const ZEROED_TEST_SEED: &Seed = &[0u8; 32];
 
             #[cfg(test)]
-            pub use ZEROED_TEST_SEED as DEFAULT_SEED;
+            pub use ZEROED_TEST_SEED as CURRENT_SEED;
 
             #[cfg(not(test))]
-            pub use crate::ephemeral_instance_keys::RANDOM_SEED as DEFAULT_SEED;
+            pub use crate::ephemeral_instance_keys::RANDOM_SEED as CURRENT_SEED;
 
             #[cfg(test)]
             mod tests {
@@ -75,7 +75,7 @@ pub mod keys {
 
                 use crate::ephemeral_instance_keys::RANDOM_SEED;
                 use crate::protocol::crypto::keys::seeds::detail::ZEROED_TEST_SEED;
-                use crate::protocol::crypto::keys::seeds::DEFAULT_SEED;
+                use crate::protocol::crypto::keys::seeds::CURRENT_SEED;
 
                 #[test]
                 fn it_should_have_a_zero_test_seed() {
@@ -84,7 +84,7 @@ pub mod keys {
 
                 #[test]
                 fn it_should_default_to_zeroed_seed_when_testing() {
-                    assert_eq!(*DEFAULT_SEED, *ZEROED_TEST_SEED);
+                    assert_eq!(*CURRENT_SEED, *ZEROED_TEST_SEED);
                 }
 
                 #[test]
