+++ src/protocol/mod.rs
@@ -1,4 +1,5 @@
 pub mod clock;
 pub mod common;
 pub mod crypto;
+pub mod info_hash;
 pub mod utils;
