+++ src/udp/request.rs
@@ -1,6 +1,6 @@
 use aquatic_udp_protocol::AnnounceRequest;
 
-use crate::protocol::common::InfoHash;
+use crate::protocol::info_hash::InfoHash;
 
 // struct AnnounceRequest {
 //     pub connection_id: i64,
@@ -17,15 +17,15 @@ use crate::protocol::common::InfoHash;
 //     pub port: Port
 // }
 
-pub struct AnnounceRequestWrapper {
+pub struct AnnounceWrapper {
     pub announce_request: AnnounceRequest,
     pub info_hash: InfoHash,
 }
 
-impl AnnounceRequestWrapper {
+impl AnnounceWrapper {
     #[must_use]
     pub fn new(announce_request: &AnnounceRequest) -> Self {
-        AnnounceRequestWrapper {
+        AnnounceWrapper {
             announce_request: announce_request.clone(),
             info_hash: InfoHash(announce_request.info_hash.0),
         }
